+++ CHANGELOG.md
@@ -5,5 +5,6 @@
 - Added functionality for irrigation
 - Added switch to enable/disable battery mode
 - Form submit replies with updated form
+- Added form to switch D4..D7 on/off manually
 
 ## 0.1.0 2020-05-31 Initial revision